requirements.txt
streamlit
sentence-transformers
PyPDF2
qdrant-client
python-dotenv
transformers
torch
google-generativeai
